ColorEngine: theme improvements/completeness
Use ThemeOverlay.Material.Light for light themes as parent: this fixes
incomplete theming when applying it to an app that has a dark theme by
default.
Also remove some of the colors and inherit them from the parent
instead, e.g. @color/white is not so good as text color for dark
themes since disabled preferences still kind of look enabled with
this, and now we have a nicer ripple color

* This commit is also included: https://github.com/AICP/frameworks_base/commit/f6f60b3763198773dd94a9ebfb17399004a9bb87

Change-Id: Ib699238f964858cd66df0c2e3a51a8d91ad88e37
Signed-off-by: Marco Marinho <marcomarinho12@gmail.com>